+++ ntex-util/src/services/mod.rs
@@ -3,8 +3,6 @@ pub mod counter;
 mod extensions;
 pub mod inflight;
 pub mod keepalive;
-pub mod sink;
-pub mod stream;
 pub mod timeout;
 pub mod variant;