+++ ntex/tests/web_ws.rs
@@ -1,6 +1,5 @@
 use std::io;
 
-use futures_util::StreamExt;
 use ntex::http::StatusCode;
 use ntex::service::{fn_factory_with_config, fn_service};
 use ntex::util::{ByteString, Bytes};
@@ -23,10 +22,9 @@ async fn service(msg: ws::Frame) -> Result<Option<ws::Message>, io::Error> {
 async fn web_ws() {
     let srv = test::server(|| {
         App::new().service(web::resource("/").route(web::to(
-            |req: HttpRequest, pl: web::types::Payload| async move {
-                ws::start::<_, _, _, web::Error>(
+            |req: HttpRequest| async move {
+                ws::start::<_, _, web::Error>(
                     req,
-                    pl,
                     fn_factory_with_config(|_| async {
                         Ok::<_, web::Error>(fn_service(service))
                     }),
@@ -71,10 +69,9 @@ async fn web_ws() {
 async fn web_ws_client() {
     let srv = test::server(|| {
         App::new().service(web::resource("/").route(web::to(
-            |req: HttpRequest, pl: web::types::Payload| async move {
-                ws::start::<_, _, _, web::Error>(
+            |req: HttpRequest| async move {
+                ws::start::<_, _, web::Error>(
                     req,
-                    pl,
                     fn_factory_with_config(|_| async {
                         Ok::<_, web::Error>(fn_service(service))
                     }),
@@ -89,33 +86,33 @@ async fn web_ws_client() {
     assert_eq!(conn.response().status(), StatusCode::SWITCHING_PROTOCOLS);
 
     let sink = conn.sink();
-    let mut rx = conn.start_default();
+    let rx = conn.receiver();
 
     sink.send(ws::Message::Text(ByteString::from_static("text")))
         .await
         .unwrap();
-    let item = rx.next().await.unwrap().unwrap();
+    let item = rx.recv().await.unwrap().unwrap();
     assert_eq!(item, ws::Frame::Text(Bytes::from_static(b"text")));
 
     sink.send(ws::Message::Binary("text".into())).await.unwrap();
-    let item = rx.next().await.unwrap().unwrap();
+    let item = rx.recv().await.unwrap().unwrap();
     assert_eq!(item, ws::Frame::Binary(Bytes::from_static(b"text")));
 
     sink.send(ws::Message::Ping("text".into())).await.unwrap();
-    let item = rx.next().await.unwrap().unwrap();
+    let item = rx.recv().await.unwrap().unwrap();
     assert_eq!(item, ws::Frame::Pong("text".to_string().into()));
 
-    let _on_disconnect = sink.on_disconnect();
+    let on_disconnect = sink.on_disconnect();
 
     sink.send(ws::Message::Close(Some(ws::CloseCode::Normal.into())))
         .await
         .unwrap();
-    let item = rx.next().await.unwrap().unwrap();
+    let item = rx.recv().await.unwrap().unwrap();
     assert_eq!(item, ws::Frame::Close(Some(ws::CloseCode::Away.into())));
 
-    let item = rx.next().await;
+    let item = rx.recv().await;
     assert!(item.is_none());
 
     // TODO fix
-    // on_disconnect.await
+    on_disconnect.await
 }